Add a basic setuptools command that prints the package name and version
This is a proof-of-concept for taking the package metadata collected by
setuptools and printing it out as a bare-bones pyproject.toml file.
Right now it only looks at the name and version (and it only detects
the static version number, without regard for setuptools_scm or anything
that would dynamically determine the version number). We'll fill in this
command class with proper detection for more metadata in the future.