Add a stub setup.py
In this commit I'm adding a setup.py file that just calls the setup()
function with no arguments. This will help support old installation
workflows, e.g. really old versions of pip, that require actually
running setup.py (as opposed to the modern workflow in which the build
backend in setuptools is capable of just reading setup.cfg directly).
But also, and perhaps more importantly for our purposes, having
a setup.py script lets us use this project on itself to demonstrate
what it does.