Add user_options to Command subclass
I determined by trial and error and a bit of reference to the setuptools
source code that subclasses of setuptools.Command need to have a class
variable user_options which defines the CLI options that the subcommand
accepts. Right now we don't have any options but there still needs to be
an empty array.

At this point the project is actually usable.