Write a README, in Markdown
In this commit I'm switching the README format from Restructured Text to
Markdown, and filling it in with some basic usage info. Some parts of
the README file (in particular the badges) will still need to be changed
later, but this at least will give some idea of what the project is
supposed to do so that people aren't entirely confused by it.