Set minimum supported Python version to 3.6
In discussion on Github,
  https://github.com/diazona/setuptools-pyproject-migration/discussions/2
we determined that we want this package to be as widely compatible as is
reasonable. Since most people are only going to use it once (or, once
per package they maintain), keeping up with the latest supported Python
release isn't that important, and if we require a newer version than
what a package maintainer is already using, that might be enough of
a barrier to discourage someone from using it.

Setting the minimum version at 3.6 allows us to use type hints (PEP 484)
and variable annotations (PEP 526), so I'm implementing that here. If we
find a reason that we need to change this, e.g. restrict to 3.7 and up
or support 3.5 for some special use case, we can always change it later.