Implement a simple test and a fixture for creating test projects
In this commit I'm adding a pytest fixture that sets up a project with
a setup.py file and other files as specified, and using that fixture in
a test that runs the pyproject conversion on a very simple project whose
setup.cfg just defines a name and version. The main point of this commit
is just to get a basic testing infrastructure set up.

In the future, I'd want to alter the plugin and test framework so that
we can actually get access to the data structure generated before it
gets written out to pyproject.toml, and we can directly make assertions
about the content of that data structure which will be more useful than
asserting the string content of a file. But that's more involved than
just the minimum required to get a test working.