Add configuration for ruff
In this commit I'm adding a configuration section for ruff that sets
the line length to 120 (same as was done for black), lists the src
directory as the package root relative to which imports should be
resolved, and sets Python 3.6 as the minimum target version.

I chose to put the configuration in pyproject.toml instead of a separate
file because it keeps the two line-length settings (for ruff and black)
close together and hopefully makes it harder to desynchronize them if we
change the line length later.