Make the documentation build successfully
In this commit I'm fixing a couple small errors that prevented
the documentation from building: I added the actual name of the project
module, and I added an Intersphinx mapping for setuptools so that sphinx
can look up documentation for setuptools classes.

The documentation still doesn't have much of anything useful but this at
least should allow the CI job to pass, then we can continue to iterate
on the documentation content itself without fear of breaking it.