Add pre-commit hook for ruff
The pre-commit hook uses version 0.0.233 of ruff, which is the last
version that accepts "py36" as the value for target-version. However
based on the comment at
  https://github.com/astral-sh/ruff/issues/2153#issuecomment-1403668601
I suspect ruff doesn't technically *support* Python 3.6, and even with
this version there's a chance it might generate code that's incompatible
with 3.6. But I don't think that's very likely. If it comes up in
the future, we can either work around it or drop support for Python 3.6
in setuptools-pyproject-migration itself. This change is more because
v0.0.234 of ruff started failing with an error if target-version is set
to anything earlier than "py37".