Skip duplicate actions when pushing in this repository
Currently when creating a pull request (at least, one not in a fork, but
at the moment there are no forks anyway), we get two copies of each
Github action, one for the branch push and one for the pull request.
The change in this commit should skip the action running for the pull
request whenever the source branch exists locally in this repository.
It shouldn't affect branches coming from other people's forks, if there
are forks in the future.

I copied the logic to do this from another project in which I worked out
how to do the same thing:
  https://github.com/pytest-dev/pytest-localserver/pull/45
The comment I've added to the workflow refers to a blog post I relied on
when figuring out how to make this work for that other project.