Split run() into two parts for generating and writing the data
This commit splits WritePyproject.run() into two methods, one that
generates the data structure corresponding to the content of
pyproject.toml and another that writes it out. This split will come in
handy for unit testing, so that we can write tests against _generate()
which don't have to go through the loop of writing out TOML-encoded
content and reading it back in.