Refactor tests to use generate() directly
In this commit I'm adding the Project.generate() method, which directly
returns the data structure generated by the plugin, without rendering it
to TOML. This gives us a more efficient way to test the plugin's output
without invoking or mimicing a subprocess and without going through
an unnecessary round of conversion to and back from TOML markup.

I duplicated all the existing tests which involve running setup.py with
script_runner.run(), except I changed the copies to use the generate()
method. We don't really need to have all these tests duplicated, but I'm
leaving it for a future commit to decide which ones from the original
batch to remove and which to keep. We'll definitely want to keep some of
the original tests that actually run setup.py, just to make sure that
the whole system does work, but going forward we can probably use
generate() to do the majority of our testing.

As part of this change, I added a call to monkeypatch.chdir() in
the project fixture so that every test which uses a project constructed
by that fixture will run in that project's root directory by default.
That's necessary because setuptools.run_setup() doesn't provide a way to
set the working directory itself like script_runner.run() does, so we
have to manually change to the right directory beforehand. I could have
added a call to chdir() as part of generate(), but changing directory as
part of a fixture allows pytest to restore the original directory after
the tests run, in case that's relevant.

In the future, perhaps we can also programmatically construct
Distribution objects instead of having to execute a setup.py file, but
that's an entirely separate feature.