Add complete TOML type information
This commit creates type aliases for the types involved in a structure
that can be rendered to TOML. This will help us use mypy to ensure that
we're not adding anything to the returned data structure that couldn't
be written out in TOML format.

In order to help convince mypy that this doesn't introduce a type
violation, I had to add the setuptools type stubs as a testing
dependency. I also added typing-extensions as a dependency to get access
to a version-independent definition of TypeAlias. (mypy does not buy it
if I just try importing TypeAlias and fall back to Type if that fails)