Add a pre-commit hook to verify Github Actions configuration
I'm using the actionlint-docker pre-commit hook to verify the workflow
files. We could use the plain old actionlint hook instead (non-Docker)
but I had some trouble compiling it on my machine so I figured this
would be easier. If anyone complains about the dependency on Docker then
we can adjust it accordingly.

This is going to help catch problems that may occur as we update the CI
configuration in upcoming commits.