Switch to more precise type definitions
This commit replaces the TOML types with a new type that precisely
defines the contents of the pyproject data structure field-by-field,
according to PEPs 518 and 621 and the documentation on the PyPA website.

This helps prevent two kinds of type checking errors. First and most
obviously, the new type is mutable - it uses List and Dict directly
instead of Sequence and Mapping - so the type checker won't complain
about assignments being made to it as we build it up.

Second, and slightly more subtle, is the fact that the type checker
didn't previously have information about the types of individual fields
in the tables, so it would apply overly strict checks to assignments at
the second level and below of the structure. For example, this affects
the line

  pyproject["project"]["dependencies"] = dependencies

In the old code, the type checker as it checks this line doesn't know
what type pyproject["project"] is, only that it's *some* type which can
be a value in a TOML table. So it will try to check every possible type
that pyproject["project"] that could be to verify that it supports
the item assignment protocol (__setitem__()) with a string key and
a List[str] value. But of course, out of all the types of values that
can exist in a TOML table, only mappings support that protocol, not
sequences or primitive values (boolean, number, string, time). The only
way to solve that, other than by giving up on typing this structure
entirely, is to give the type checker information about which types of
value is supposed to be in each field. Hence the use of TypedDict.

And of course, this is going to make it easier to catch erroneous
assignments to fields of the pyproject data structure in the future.