Drop down to two tests that actually run the setup.py command
This commit changes the set of tests which actually run
`setup.py pyproject` to be just two: one extremely minimal test which
just has a name and version, and another comprehensive test that is
supposed to include all relevant fields that can appear in setup.cfg.
The idea is that as we add support for more fields, we'll adjust
the expected output of that test to keep up. I populated the dummy
setup.cfg file for the latter test with a bunch of fields that exist
in setuptools' own setup.cfg file as well as this project's setup.cfg.