Remove release environment from tox and Github Actions
I want to force all releases to go through Github for traceability, so
I'm removing the release environment from tox.ini to make it difficult
for someone to create a release locally. And since the existing release
job in Github Actions uses the tox environment, I'm removing that job as
well. In an upcoming commit I'm going to add a brand new GHA workflow to
produce a release.