Add CI tests on Python back to 3.6, and expand pre-release testing
Prior to this point, we had both pre-/post-merge testing and pre-release
testing set up to run on a few Python versions since 3.8, across
multiple operating systems. But that configuration didn't make the most
sense, for a few reasons:

- This package is basically OS-independent, and it's unlikely that we're
  going to find bugs on one OS that don't occur on all other operating
  systems.
- Similarly, very little in this package depends on even the minor
  version of Python, and so there are probably not going to be a whole
  lot of bugs that occur on some versions but not others - and of those,
  most of them will probably occur on all versions less than 3.X or all
  versions greater than 3.X for some X. So we can usually get away with
  testing the oldest and newest versions.
- For the same reason, we probably don't need extensive testing on PyPy.

With this in mind, I trimmed down our pre-/post-merge tests to just five
configurations:

- The oldest version of Python the package supports
- The oldest stable version of Python
- The newest stable version of Python
- Any reasonably close-to-stable pre-release version of Python
- And the newest stable version of Python with dependencies pinned to
  the oldest versions we claim to be compatible with

All are on only Linux. I think this set of configurations should have
a good chance of catching most bugs.

However, we do want to be confident about the stability of our releases,
so I changed the pre-release testing to cover all Python versions we
claim to support (including PyPy), all operating systems on a variety of
Python versions, as well as testing with pinned old dependency versions
on both old and new Python versions. This should make it pretty unlikely
that any system-dependent bugs make it into a release.