Limit Python versions of some dependencies
Some of the tests on Python 3.6 fail because there are no versions of
certain dependencies available on that version of Python. However,
the dependencies affected are only for testing and static analysis, not
dependencies of the package itself, so we can just drop those
dependencies on 3.6 and skip any related tests.

In the future I do want to break out the static analysis into separate
tox environments and only run it on one version of Python to save work.