Allow old method of passing command to ScriptRunner.run()
The run() method of the ScriptRunner class from pytest-console-scripts
used to only accept the command as individual positional arguments, not
as a list. This changed in pytest-console-scripts 1.4, and the old
behavior of accepting positional arguments is still supported but is
deprecated. I wanted to keep using a list when possible to set this
package up for when we can drop support for older versions of
pytest-console-scripts (which will coincide with dropping support for
Python 3.6), but in order to maintain compatibility with Python 3.6 for
now, I'm adding some logic to detect when pytest-console-scripts is
an older version and pass the command as positional arguments.

In order to do that, I had to introduce a couple new imports:
the packaging package (for a version number parser), and
importlib.metadata or importlib_metadata (for the ability to check
the version of pytest-console-scripts). Both these packages are
dependencies of other things we already use for testing, so this is
really just changing indirect dependencies to direct ones. We can get
rid of all this once we drop support for Python 3.6.