Add support for translating authors and maintainers to pyproject.toml
This implementation assumes that the author, author_email, maintainer,
and maintainer_email strings are comma-separated lists of authors or
maintainers respectively. A number of projects (including this one) use
that method to list multiple authors/maintainers (as per
e.g. https://stackoverflow.com/a/10005265), and for projects that list
only a single person in the field, this won't hurt. We can expand this
logic in the future to support other methods of specifying multiple
authors/maintainers if necessary, perhaps even to the point of adding
a command-line option to inform the plugin how to parse those fields.