Add tests of the plugin with author and maintainer metadata
These tests run various combinations of author and maintainer metadata
through the plugin. The tests only verify the "typical" cases; I'm
leaving it to the unit tests for _transform_contributors() to ensure
that weird cases are handled in the proper way.

I made a judgment call to treat any missing metadata as an empty string,
but to never generate an author or maintainer entry in pyproject.toml
unless at least one of the name or email address is nonempty. This was
an arbitrary choice, at least in the sense that I didn't do any research
to determine how those cases are supposed to be handled (if there is any
standard for them). If necessary we can fix up the behavior later.

I wrote these tests to use the project fixture rather than manually
constructing Distribution objects because if I try the latter way,
it looks like setuptools_scm takes over and computes the version number
from the actual setuptools-pyproject-migration project, rather than
using the hard-coded version 0.0.1, which causes the tests to fail.
Hopefully in a future commit we can fix that.