Add Markdown linter as a pre-commit hook
I examined a few different Markdown linters that supply pre-commit hooks
to find one that would work for our project. The one by Igor Shubovych
seems to be popular and actively developed, supports configuration, and
offers auto-fixing of some issues, so I think it's a good choice for
this project.