Add support for classifiers
I chose to treat an empty list of classifiers the same as the absence of
any classifiers because it looks like setuptools/distutils doesn't make
that distinction: even if the list of classifiers exists but is empty,
`Distribution.get_classifiers()` returns `None`. (This will be a common
pattern with list-valued fields, because distutils uses common code for
parsing all those fields.)

This means that if someone has a setup.cfg file with an empty list of
classifiers, like this:

[metadata]
classifiers =

that empty list will not appear in pyproject.toml. This should be easy
enough for any users to deal with. (And if really necessary, we might
be able to patch setuptools/distutils to distinguish between empty and
absent lists, but I don't want to put in that effort without knowing
that it's going to be useful.)