Ignore a warning about deprecated distutils
A warning about the deprecation of distutils in Python 3.12 appears in
CI when running under Python 3.11 with old versions of our dependencies.
(Most likely it's related to using old setuptools, but I can't reproduce
it locally so I can't be sure.) This warning isn't relevant to us
because we can import the vendored version of distutils from setuptools
itself, so I'm setting the warning filter up to ignore it.