Add support for description
This one is mostly pretty simple, we can just copy the description field
from the Distribution object.

However, the handling of an empty or absent description gets a little
complicated. It appears that older versions of setuptools (specifically
I tested 56.0.0) represent either an empty description or the absence of
a description field as "UNKNOWN", but recent versions (e.g. 68.0.0) will
distinguish between the absence of a description field (represented as
None) and an empty description field (represented as an empty string).
To keep things as consistent as possible, I've opted not to distinguish
between absent and empty descriptions here; either way, no description
field will be generated in the resulting pyproject.toml file. But in
the future, if we find a use case where it makes sense to distinguish
between absent and empty description, I think it'd be totally reasonable
to change the code to do so for versions of setuptools that allow it,
and just issue a warning for older versions of setuptools that don't.