Add support for the readme field
The readme field in pyproject.toml is populated from long_description
and long_description_content_type (if present) in the setuptools
configuration. This one is a little complicated though, because calling
Distribution.get_long_description() will just return the text content of
the long description, whereas what we need to write into pyproject.toml
is the name of a file containing (what setuptools calls) the long
description - in other words, the name of the readme file.

What I've done here to handle that is dig into the more obscure fields
of the Distribution object to find where it stores the original, raw
value of the long_description field, and if that field starts with
`file:`, I extract the filename and copy it over into pyproject.toml.
On the other hand, if the long_description is a string, I've made it
write that string out to a file and put the filename in pyproject.toml.
The file is named with an extension corresponding to the content type
if available; I introduced a bit of custom code to find the extension
for a given content type to work around a bug in older versions of
Python.

This makes the plugin do something that has not been done for any
previous field, namely creating a brand new file that needs to become
part of the project being converted. I'm not sure how best to handle
this. The current implementation is fine to get something that basically
works, but it might be better to issue some kind of a warning or
notification to the user that a new file was created, or perhaps exit
with an error that prompts the user to explicitly specify that it's okay
to write a new file with a command-line option. (Maybe that option could
accept the new filename as an argument.) We should probably revisit this
in the future based on what seems to be most useful.

The content type is a bit more straightforward: I can just copy it from
setuptools into the pyproject data structure, if it's there, and if not
I just use a string value for the `readme` field. I have not made any
effort to have the code guess the content type based on the content of
the long description. That's something we could potentially add in
the future, but it's going beyond what I have in mind as the purpose of
this plugin, which is to copy information from setuptools to
pyproject.toml as faithfully as possible.