Update test_everything() with keywords
This commit adds the expected result for the keywords field to
test_everything().

I changed one thing in the input setup.cfg file: the list of keywords
was space-separated, but it should have been comma-separated. In fact
the core metadata specification was changed from using spaces to commas
precisely because that's what distutils and setuptools did.
  https://packaging.python.org/en/latest/specifications/core-metadata/#core-metadata-keywords