Streamline the manual creation of Distribution objects with new fixtures
When testing WritePyproject, we need to give it a Distribution object,
which can be created either by calling setuptools.setup() or by manually
calling the constructor and passing in a dict of attributes. The latter
way is faster to run because it doesn't have to create and read a file,
but it's a little unwieldy to write that code over and over again. Plus,
we've had problems where the Distribution object will read
setuptools configuration files in whatever is the current working
directory when the test runs, and will set some of its attributes based
on that configuration. (Specifically, it seems to use setuptools_scm to
set the version number, but maybe this would apply to other dynamic
attributes, I'm not sure.)

This commit adds a fixture, make_write_pyproject, that will address
the problem in two ways:

- It ensures that any test requesting the fixture will run in an empty
  working directory
- It also provides a convenience function that creates a Distribution
  from the given keyword arguments and then creates a WritePyproject
  from that Distribution.

We can now use this fixture to avoid having to run the full setuptools
configuration procedure if we don't have to.