Change the fixture to set name and version by default
In this commit, I'm changing the make_write_pyproject fixture to set
a default name and version if those are not passed as keyword arguments.
Those are the only two attributes required by the Distribution
constructor, but we often don't care what their values are. If a test
does care, the values are accessible as class attributes DEFAULT_NAME
and DEFAULT_VERSION.

As a side effect of making the fixture return a class, I had to drop
typing information, because names in conftest.py aren't made available
to test code except through fixtures. It should be possible to bring
back the typing information by creating an extra package in the main
project for test helper code, or by using something like
pytest-helpers-namespace, but I don't think that's strictly necessary
to get this code merged.