Add support for the requires-python field
This commit adds code to the plugin to set requires-python based on
the value of python_requires in the Distribution. It's pretty
straightforward because both fields have the same kind of value:
a PEP 440-compliant version string. I've chosen to take the value from
the parsed version of the field, rather than digging into the metadata
to try to get the original unparsed field value, because I prefer to
stick to the mainstream API of setuptools whenever possible, and it
should produce an equivalent result either way: using the parsed version
means that the string added to pyproject.toml might not be exactly
the same as the one that was given to setuptools, but it will select
the same Python versions.