Add tests of requires-python
These tests do the obvious thing of confirming that the specifier string
appearing in the plugin's output selects the same Python versions as
the string originally given to the Distribution. Following how I
implemented the code that handles requires-python, I've made the tests
ensure that the input and output select the same Python versions, but
not required that the string is passed through unmodified.

In one of the doc comments I added a reference to the packaging library,
so I updated the Sphinx configuration to include it. Even though we
don't actually generate documentation for test files, I think it's nice
to keep the configuration up to date so we could.