Add support for the urls field
In this commit I'm copying over the project_urls field from setuptools
into the urls field of pyproject.toml. The two fields follow the same
format so it's a straightforward assignment.

setuptools also recognizes individual fields for some specific URLs:

- url: the URL of the project homepage
- download_url: the URL of the downloadable code

The implementation I have here doesn't do anything special with these
URLs, which is mainly to be consistent with the philosophy used in
the rest of the plugin, that we should avoid making assumptions and
just faithfully translate information from setuptools to pyproject.toml.
But this is a case where it probably makes sense to take a different
approach: there are surely many projects that only have url and/or
download_url set, but nothing in project_urls, and they would expect
those URLs not to be discarded. I'm not sure what is the right way to
handle those fields but we should probably add something for them in
a future commit.