Add tests of the urls field
Here I'm adding tests of the various ways in which we might set
the urls field in pyproject.toml. This includes tests of the individual
URL fields recognized by setuptools, even though the code doesn't
currently add them to pyproject.toml, because it's important to specify
the plugin's behavior with respect to those fields.