Add a console script to run the pyproject command
In this commit I'm adding a console script that will do the equivalent
of running `python setup.py pyproject`, including faking a stub setup.py
file if one doesn't exist. This will let people install the project with
pipx and have a simple shell command that will generate their
pyproject.toml, so they don't have to go to the trouble of setting up
a virtualenv manually. It also means projects that have a setup.cfg file
but no setup.py will be able to use this project.