Generalize setup command tests to use the console script too
This commit changes a couple of the tests where we currently run
setup.py pyproject and parametrizes them such that they will also run
the console script with and without a stub setup.py file. The output
is expected to be the same either way.

We can apply this same generalization to other tests in the future as
needed, but it will require moving the new fixture to conftest.py (which
is fine, there's no reason not to do that, but we might want to think
more carefully about its name at that point).