Add support for optional dependencies
In this commit I'm adding support for optional dependencies in
pyproject.toml, populated by the extras_require field in setuptools.
I've made the implementation sort and deduplicate the dependencies for
each extra, both for consistency with how we handle non-optional
dependencies, and because setuptools exposes optional dependencies with
environment markers in separate lists per marker, so the original order
of the dependencies is not available. (It could be reconstructed by
digging into the DistributionMetadata object, but that would be a lot of
work for a fairly useless change.) However, extras with no dependencies
are still preserved.