Add tests of optional dependencies
This set of tests tries to cover all the different features that can
occur in dependency specifications, including version constraints (like
"spam>=1.5,<2") and environment markers ("spam;python_version<'3.10'"),

When setuptools reads in dependency specifications, it normalizes their
formatting and the ordering of different constraints, so the actual
dependency spec that gets written to pyproject.toml won't necessarily be
a string match to the original one as written in setup.cfg or setup.py.
I've written the tests to take that normalization into account by hand-
coding the expected output in each test where normalization would make
it differ from the input. Alternatively, I could have used pep508_parser
to parse the actual and expected dependency specifications so that
formatting differences wouldn't be considered when checking the output,
but the parsed result is a non-hashable type (a list) which makes it
a little trickier to deduplicate dependencies in the way that setuptools
does, so I didn't bother with that. It would certainly be doable with
a bit of work if we find the need to in the future.