Add a docs page about the motivation for this project
This is meant to give a quick overview of why this project is useful and
where it fits into the Python packaging ecosystem. It's not all that
well written, I think, so I will probably want to come back and revise
this new page later (or someone else can), but I'll probably try to get
some input from other people about what would be useful to include.