Add WIP repo status badge
This badge text comes from https://www.repostatus.org, and appears to be
a fairly well-defined way to indicate the status of a project. It's
similar to the "Development Status" trove classifiers, but is more
visible and not specific to Python.

This is not super important but seems like a nice bit of added value.