Add testing of real distribution packages
This commit adds a framework that makes it easy to test the plugin on
real Python packages. We can list the name and version of a package
available on PyPI and the framework will download and extract the source
code, run the plugin to generate the content of pyproject.toml, and
check that the generated content gives the same metadata as the actual
published metadata for the package.

Packages that are good candidates for this testing framework are those
which have a setup.py or setup.cfg file but not a pyproject.toml file,
at least not one that contains core metadata. I'm not sure offhand of
any great examples, but I included two packages as proofs-of-concept:
pytest 7.3 and pytest-localserver 0.8. These are packages I happen to be
particularly familiar with but otherwise they're arbitrarily chosen.
Neither of these tests pass right now, so I've marked them as expected
failures. We'll have to fix those issues in the future.

I've marked the new tests as slow so that they don't run during normal
pre-merge/post-merge testing.