Move test support code to a separate package
In this commit I'm introducing a new top-level directory to hold helper
code used during testing. We don't need to put all helper code here,
but I'd like to use it for anything substantial, to avoid cluttering up
conftest.py files. We can also use it for code that needs to be imported
from multiple test files in different directories, like helper classes
that are returned from fixtures (because we need them in the test code
for type annotations).

I added the pythonpath configuration option in pytest.ini to put the new
test support directory on sys.path when running tests. This option
requires pytest>=7, so I've updated the dependency constraint in
setup.cfg to match.

Right now, the Project class qualifies on both counts, and
WritePyprojectFactory is also something that needs to be importable from
different test files, so I've moved both. In upcoming commits, I'm going
to be introducing a large body of test support code for running tests on
live packages, which will also go in the new test support directory.