Simplify distribution version detection
In this commit I'm changing the test support code to just import either
importlib_metadata or importlib.metadata, rather than importing only
the version function from it. I've also created a little function to do
the version comparison we need, mainly because it makes the code a bit
cleaner than hard-coding that check. It is a bit of overkill to have
a function that's only used in one place, but I think the readability
benefits are worth it, and who knows, that function might be useful
elsewhere in the future.

I'm doing this because in the live project testing code I'm going to add
later, I'll need access to other things from importlib_metadata or
importlib.metadata and I want to only make the decision about which one
to import in one place.