Add documentation of test support API
This commit introduces documentation of the code in the test_support
directory, which will be useful to anyone working on the tests. I added
docstrings for a few things which were previously missing them in order
to reduce the number of warnings emitted by sphinx.

I wasn't able to get rid of all the warnings, though, so I had to remove
the -W option (upgrades warnings to errors) from the sphinx invocation
in tox.ini. At some point in the future we should fix those warnings and
restore the -W option. (I'm sure they must be fixable, it's just
difficult to figure out how.)